Stage commit

+++ src/client/app/services/mail.service.js
@@ -5,10 +5,14 @@
         .module('app.services')
         .factory('mail', mail);
 
-    mail.$inject = ['CONFIG', '$resource', '$http'];
+    mail.$inject = ['CONFIG', '$resource', '$http', '$rootScope', 'Upload'];
 
-    function mail(CONFIG, $resource, $http) {
+    function mail(CONFIG, $resource, $http, $rootScope, Upload) {
         var API_URL = CONFIG.APIHost + '/mail';
+
+        var answerData = {};
+
+        var fwdData = {};
 
         var resource = $resource(API_URL,
             {},
@@ -55,10 +59,15 @@
             }
         );
 
-        // var messages = [];
-
         function post(params, data) {
-            return resource.post(params, data).$promise;
+            return resource.post(params, data).$promise
+                .then(function (response) {
+                    console.log('params', data);
+                    if (data.cmd === 'send') {
+                        $rootScope.$broadcast('mail:send:success');
+                    }
+                    return response;
+                });
         }
 
         function get(params, data) {
@@ -69,12 +78,53 @@
             return resource.put(params, data).$promise;
         }
 
+        function upload(params, data, files) {
+            var foramttedData = {
+                to: data.to,
+                body: data.body,
+                cmd: 'add-attach'
+            };
+
+            _.forEach(files, function (file, i){
+                var name = 'file' + i;
+                foramttedData[name] = file;
+            });
+
+            return Upload.upload({
+                url: CONFIG.APIHost + '/mails/add-attach',
+                data: foramttedData
+            });
+        }
+
         function getById(params, data) {
             return resource.getById(params, data).$promise;
         }
 
         function move(params, data) {
             return resource.move(params, data).$promise;
+        }
+
+        function moveToFolder(folder, data) {
+            var messages = angular.copy(data);
+
+            if (messages.isLoading || !messages.checked.length) return;
+
+            move({}, {
+                messages: messages.checked,
+                mboxnew: folder.name
+            }).then(function () {
+                $rootScope.$broadcast('mailBox:sync');
+            });
+
+            _.forEach(messages.checked, function (checked) {
+                _.remove(messages.items, function (message) {
+                    return message.number === checked.number;
+                });
+            });
+
+            messages.checked = [];
+
+            return messages;
         }
 
         function flag(params, data) {
@@ -85,16 +135,25 @@
             return resource.deflag(params, data).$promise;
         }
 
-        function destroy(params, data) {
-            return $http({
-                url: API_URL + '/' + data.id,
+        function destroy(data) {
+            var messages = angular.copy(data);
+
+            if (messages.isLoading || !messages.checked.length) return;
+
+            $http({
+                url: API_URL + '/' + 1,
                 method: 'DELETE',
-                data: data,
+                data: {
+                    messages: messages.checked
+                },
                 headers: {
                     "Content-Type": "application/json;charset=utf-8"
                 }
             });
-            // return resource.destroy(params, data).$promise;
+
+            messages.checked = [];
+
+            return messages;
         }
 
         function setSeen(data) {
@@ -102,10 +161,17 @@
 
             if (messages.isLoading || !messages.checked.length) return;
 
-            var ids = [];
+            messages.isLoading = true;
+
+            flag({}, {
+                messages: messages.checked,
+                flag: 'Seen'
+            }).then(function (response) {
+                messages.isLoading = false;
+                $rootScope.$broadcast('mailBox:sync');
+            });
 
             _.forEach(messages.checked, function (checked) {
-                ids.push(checked.number);
                 _.forEach(messages.items, function (item) {
                     if (checked.number == item.number) {
                         item.seen = true;
@@ -115,23 +181,6 @@
 
             messages.checked = [];
 
-            _.forEach(messages.items, function (item) {
-                _.forEach(ids, function (id) {
-                    if (item.number === id) {
-                        messages.checked.push(item);
-                    }
-                });
-            });
-
-            messages.isLoading = true;
-
-            flag({}, {
-                ids: ids,
-                flag: 'Seen'
-            }).then(function (response) {
-                messages.isLoading = false;
-            });
-
             return messages;
         }
 
@@ -140,10 +189,17 @@
 
             if (messages.isLoading || !messages.checked.length) return;
 
-            var ids = [];
+            messages.isLoading = true;
+
+            deflag({}, {
+                messages: messages.checked,
+                flag: 'Seen'
+            }).then(function (response) {
+                messages.isLoading = false;
+                $rootScope.$broadcast('mailBox:sync');
+            });
 
             _.forEach(messages.checked, function (checked) {
-                ids.push(checked.number);
                 _.forEach(messages.items, function (item) {
                     if (checked.number == item.number) {
                         item.seen = false;
@@ -153,23 +209,6 @@
 
             messages.checked = [];
 
-            _.forEach(messages.items, function (item) {
-                _.forEach(ids, function (id) {
-                    if (item.number === id) {
-                        messages.checked.push(item);
-                    }
-                });
-            });
-
-            messages.isLoading = true;
-
-            deflag({}, {
-                ids: ids,
-                flag: 'Seen'
-            }).then(function (response) {
-                messages.isLoading = false;
-            });
-
             return messages;
         }
 
@@ -178,10 +217,16 @@
 
             if (messages.isLoading || !messages.checked.length) return;
 
-            var ids = [];
+            messages.isLoading = true;
+
+            flag({}, {
+                messages: messages.checked,
+                flag: 'Flagged'
+            }).then(function (response) {
+                messages.isLoading = false;
+            });
 
             _.forEach(messages.checked, function (checked) {
-                ids.push(checked.number);
                 _.forEach(messages.items, function (item) {
                     if (checked.number == item.number) {
                         item.important = true;
@@ -191,23 +236,6 @@
 
             messages.checked = [];
 
-            _.forEach(messages.items, function (item) {
-                _.forEach(ids, function (id) {
-                    if (item.number === id) {
-                        messages.checked.push(item);
-                    }
-                });
-            });
-
-            messages.isLoading = true;
-
-            flag({}, {
-                ids: ids,
-                flag: 'Flagged'
-            }).then(function (response) {
-                messages.isLoading = false;
-            });
-
             return messages;
         }
 
@@ -216,10 +244,16 @@
 
             if (messages.isLoading || !messages.checked.length) return;
 
-            var ids = [];
+            messages.isLoading = true;
+
+            deflag({}, {
+                messages: messages.checked,
+                flag: 'Flagged'
+            }).then(function (response) {
+                messages.isLoading = false;
+            });
 
             _.forEach(messages.checked, function (checked) {
-                ids.push(checked.number);
                 _.forEach(messages.items, function (item) {
                     if (checked.number == item.number) {
                         item.important = false;
@@ -229,24 +263,23 @@
 
             messages.checked = [];
 
-            _.forEach(messages.items, function (item) {
-                _.forEach(ids, function (id) {
-                    if (item.number === id) {
-                        messages.checked.push(item);
-                    }
-                });
-            });
-
-            messages.isLoading = true;
-
-            deflag({}, {
-                ids: ids,
-                flag: 'Flagged'
-            }).then(function (response) {
-                messages.isLoading = false;
-            });
-
             return messages;
+        }
+
+        function setAnswerData(data) {
+            answerData = data;
+        }
+
+        function getAnswerData() {
+            return answerData;
+        }
+
+        function setFwdData(data) {
+            fwdData = data;
+        }
+
+        function getFwdData() {
+            return fwdData;
         }
 
         return {
@@ -261,7 +294,13 @@
             setSeen: setSeen,
             setUnSeen: setUnSeen,
             setImportant: setImportant,
-            setUnImportant: setUnImportant
+            setUnImportant: setUnImportant,
+            moveToFolder: moveToFolder,
+            getAnswerData: getAnswerData,
+            setAnswerData: setAnswerData,
+            upload: upload,
+            setFwdData: setFwdData,
+            getFwdData: getFwdData
         }
     }
